Switched comment to `wt`, fixed bashism & tabbing
Most of the `read` commands were formatted according to bash-specific
syntax, so I updated this to resolve that issue.

The RSS feed previously relied on the `gen_page` comment, but I changed
it to `wt` in order to match with the rest of the scripts. If you
previously managed an RSS feed using these scripts, run:

  sed -i 's/<!-- gen_page -->/<!-- wt -->/' production/rss.xml

Previously, the RSS feed added tabs to all text, regardless of whether
it was preformatted or not. `push_post` now selectively adds tabs to
lines, based on whether they are between a <pre>...</pre> tag or not.

`./setup_rss` now updates the `$BASE_URL` variable under `./push_post`